[expo-updates][ios] Convert updates logger tests to quick/nimble
